requirements.txt
aiofiles==24.1.0
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.3.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.2
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.1
orjson==3.10.3
pillow==11.0.0
platformdirs==4.2.2
pyasn1==0.6.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
stripe==11.4.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.29.0
uvloop==0.21.0
validator==0.7.1
virtualenv==20.26.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
